// models/enum/codeList.ts
export enum CodeList {
    success = 0,
    failure = 1,

    hop_again = 11,


    browser_not_found = 10,
    browser_not_allowed = 11,
    babel_cookie_missing = 15,

    chapter_checked = 50,
    chapter_created = 51,
    chapter_updated = 52,
    chapter_creation_failed = 53,
    chapter_update_failed = 54,
    chapter_already_parsed = 55,
    chapter_premium = 56,
    chapter_invalid = 57,

    novel_checked = 80,
    novel_created = 81,
    novel_updated = 82, 
    novel_creation_failed = 83,
    novel_update_failed = 84,
    novel_fetch_checked = 90,
    novel_fetch_created = 91,
    novel_fetch_updated = 92, 
    novel_fetch_code_wrong = 93,
    novel_fetch_update_failed = 94,
    novel_fetch_error = 95

    
}